% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/103-AppRunner.R
\name{AppRunner}
\alias{AppRunner}
\alias{runImageScrambler}
\title{Runs Image Analysis Applications}
\usage{
runImageScrambler(input_image, output_image)
}
\description{
Runs Image Analysis Applications

Depending on session context, run as a command-line program
}
