% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-ImageProcessor.R
\docType{class}
\name{ImageProcessor}
\alias{ImageProcessor}
\title{A SparkR application that leverages Spark-RAPIDS to perform GPU-accelerated
various operations over provided image - Scrambling, Grayscaling}
\arguments{
\item{input_image}{Path to the input image file}

\item{output_image}{Path to the GrayScale output image file}

\item{mode}{0 = grayscale, 1 = scrambler}

\item{img}{Numeric array of the input image}

\item{df}{Spark DataFrame with r, g, b channels}

\item{res}{Local DataFrame containing image pixel values}

\item{h}{Image height}

\item{w}{Image width}
}
\value{
Initialized ImageProcessor instance

Numeric array representing the input image

Spark DataFrame with columns: row, col, r, g, b

Local DataFrame with scrambled pixels

Local R DataFrame with GrayScale values
}
\description{
Fields:
\itemize{
\item input_image  : character, path to the input image file
\item output_image : character, path where the output scrambled image is written
\item mode         : 0 = scrambles image, 1 grayscales the image
\item clazz        : character, fully qualified class identifier for logging
}

Includes structured logging and error handling.
}
\details{
Methods:
\itemize{
\item initialize()
\item startSpark(),
\item loadImage(),
\item toDataFrame(),
\item registerGrayScaleUDF(),
\item scrambleImage(),
\item computeGrayScale(),
\item saveImage(),
\item run()
}
}

