% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-ImageStacker.R
\docType{class}
\name{ImageStacker}
\alias{ImageStacker}
\title{A SparkR application that leverages Spark-RAPIDS to perform GPU-accelerated
image stacking.}
\arguments{
\item{output_image}{Path to the generated image file}

\item{width}{}

\item{height}{}

\item{path:}{fully qualified file path for image}

\item{parquetPath:}{fully qualified file path for input parquet}

\item{outputImage:}{fully qualified file path for output image}

\item{redExpr}{: expression for red column}

\item{greenExpr}{: expression for green column}

\item{blueExpr}{: expression for blue column}

\item{parquetPath}{: fully qualified file path for input parquet}

\item{outputImage}{: fully qualified file path for output image}
}
\value{
Initialized ImageStacker instance

width
}
\description{
Fields:
\itemize{
\item output_image : character, path where the images are stacked to
\item width        : integer, image width
\item height       : integer, image height
\item clazz        : character, fully qualified class identifier for logging
\item logger       : logger class
}

Examples:
\itemize{
\item Solid blue
img$colorize("0", "0", "255")
}
}
\details{
Methods:
\itemize{
\item initialize()
}

\itemize{
\item Horizontal gradient (red increases with x)
img$colorize("CAST(x * 255 / 3000 AS INT)", "0", "0")
\item Diagonal green-blue blend
}
}

