% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-Logger.R
\docType{class}
\name{Logger-class}
\alias{Logger-class}
\alias{Logger}
\title{Logging class}
\description{
Provides a simple logging utility using futile.logger
}
\details{
Methods:
\itemize{
\item info
\item warn
\item error
}
}

