% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-AppRunner.R
\name{AppRunner}
\alias{AppRunner}
\alias{runImageProcessor}
\alias{runImageStacker}
\title{Runs ImageProcessor application}
\usage{
runImageProcessor(input_image, output_image, mode = 0, height = 700)

runImageStacker(
  redExpr = "255",
  greenExpr = "0",
  blueExpr = "0",
  parquetPath = NA,
  imagePath = NA,
  height = 350,
  width = 625
)
}
\description{
Runs ImageProcessor application

Runs Image Stacker application

Depending on session context, run as a command-line program
}
