% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-AppRunner.R
\name{AppRunner}
\alias{AppRunner}
\alias{runImageScrambler}
\alias{runImageStacker}
\title{Runs Image Scrambler application}
\usage{
runImageScrambler(input_image, output_image)

runImageStacker(
  redExpr = "255",
  greenExpr = "0",
  blueExpr = "0",
  parquetPath = NA,
  imagePath = NA,
  height = 0,
  width = 0
)
}
\description{
Runs Image Scrambler application

Runs Image Analysis Applications

Depending on session context, run as a command-line program
}
