% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-AppRunner.R
\name{AppRunner}
\alias{AppRunner}
\alias{runApp}
\title{Runs GrayScale Application}
\usage{
runApp(input_image, output_image)
}
\description{
Runs GrayScale Application

Depending on session context, run as a command-line program
}
